add --tag-only flag

+++ scripts/release.sh
@@ -12,7 +12,8 @@ if [ -z "${BRAD_PITT_HOME}" ]; then
 	exit 1
 fi
 
-# This block discovers the command line flag `--dry-run`
+# This block discovers the command line flags
+# `--dry-run` and `--tag-only`
 # and passes on positional arguments as $1, $2, etc.
 if [[ $# > 0 ]]; then
     DRYRUN=
@@ -22,6 +23,10 @@ if [[ $# > 0 ]]; then
         case $key in
             --dry-run)
             DRYRUN="--dry-run"
+            shift
+            ;;
+		    --tag-only)
+            TAGONLY="--tag-only"
             shift
             ;;
             *)
@@ -49,11 +54,19 @@ if [[ $# != 2 ]]; then
     echo " - source_branch is a git branch available on the remote that will be promoted"
     echo " - dest_branch is the git branch that will point to the same commit that source_branch points to"
     echo
-    echo "Usage: $(basename $0) source_branch dest_branch [--dry-run]"
+    echo "Usage: $(basename $0) source_branch dest_branch [--dry-run] [--tag-only]"
+    echo
+    echo "    --dry-run			Prints out but does not run commands"
+	echo
+	echo "    --tag-only        Creates the tag for the new version, but"
+	echo "                      does not promote the source branch to the"
+	echo "                      destination branch, or change the head of the"
+	echo "                      destination branch in any way."
     echo
     echo "Example: $(basename $0) release/v1.0 main"
 	echo "          (this will force the main branch to point at the release/v1.0 branch)"
 	echo "          (that way, cloning the repo will clone the default (main) branch, which points at v1.0)"
+    echo
     exit 1
 fi
 
@@ -94,17 +107,25 @@ if [[ ${DRYRUN:-} == true ]]; then
 	echo
 	echo "git fetch --all"
 	echo "git -c advice.detachedHead=false checkout ${REMOTE}/$PROMOTE_FROM_BRANCH"
-	echo "git checkout -B $PROMOTE_DEST_BRANCH"
+	if [[ ${TAGONLY:-} == false ]]; then
+		echo "git checkout -B $PROMOTE_DEST_BRANCH"
+	fi
 	echo "git tag $RELEASE_TAG"
-	echo "git push --force $REMOTE $PROMOTE_DEST_BRANCH"
+	if [[ ${TAGONLY:-} == false ]]; then
+		echo "git push --force $REMOTE $PROMOTE_DEST_BRANCH"
+	fi
 	echo "git push --tags $REMOTE"
 	echo
 else
 	git fetch --all
 	git -c advice.detachedHead=false checkout ${REMOTE}/$PROMOTE_FROM_BRANCH
-	git checkout -B $PROMOTE_DEST_BRANCH
+	if [[ ${TAGONLY:-} == false ]]; then
+		git checkout -B $PROMOTE_DEST_BRANCH
+	fi
 	git tag $RELEASE_TAG
-	git push --force $REMOTE $PROMOTE_DEST_BRANCH
+	if [[ ${TAGONLY:-} == false ]]; then
+		git push --force $REMOTE $PROMOTE_DEST_BRANCH
+	fi
 	git push --tags $REMOTE
 fi
 echo "Done. Success. Thank you. Goodbye."
